+++ mlflow/server/js/src/components/ExperimentListView.js
@@ -10,8 +10,7 @@ import { Link } from 'react-router-dom';
 class ExperimentListView extends Component {
   static propTypes = {
     onClickListExperiments: PropTypes.func.isRequired,
-    // If activeExperimentId is undefined, then the active experiment is the first one.
-    activeExperimentId: PropTypes.number,
+    activeExperimentId: PropTypes.number.isRequired,
     experiments: PropTypes.arrayOf(Experiment).isRequired,
   };
 
@@ -45,13 +44,8 @@ class ExperimentListView extends Component {
                className="collapser fa fa-chevron-left login-icon"/>
           </div>
           <div className="experiment-list-container" style={{ height: experimentListHeight }}>
-            {this.props.experiments.map((e, idx) => {
-              let active;
-              if (this.props.activeExperimentId) {
-                active = parseInt(e.getExperimentId(), 10) === this.props.activeExperimentId;
-              } else {
-                active = idx === 0;
-              }
+            {this.props.experiments.map((e) => {
+              const active = parseInt(e.getExperimentId(), 10) === this.props.activeExperimentId;
               let className = "experiment-list-item";
               if (active) {
                 className = `${className} active-experiment-list-item`;
