+++ mlflow/server/js/src/components/HomePage.js
@@ -1,18 +1,26 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
+import ExperimentPage from './ExperimentPage';
 import { connect } from 'react-redux';
 import { getUUID, listExperimentsApi } from '../Actions';
 import RequestStateWrapper from './RequestStateWrapper';
 import './HomePage.css';
-import HomeView from './HomeView';
+import ExperimentListView from './ExperimentListView';
 
 class HomePage extends Component {
+  constructor(props) {
+    super(props);
+    this.onClickListExperiments = this.onClickListExperiments.bind(this);
+  }
+
   static propTypes = {
+    match: PropTypes.object.isRequired,
     dispatchListExperimentsApi: PropTypes.func.isRequired,
-    experimentId: PropTypes.number,
+    experimentId: PropTypes.number.isRequired,
   };
 
   state = {
+    listExperimentsExpanded: true,
     listExperimentsRequestId: getUUID(),
   };
 
@@ -20,19 +28,51 @@ class HomePage extends Component {
     this.props.dispatchListExperimentsApi(this.state.listExperimentsRequestId);
   }
 
+  onClickListExperiments() {
+    this.setState({ listExperimentsExpanded: !this.state.listExperimentsExpanded });
+  }
+
   render() {
-    return (
-      <RequestStateWrapper requestIds={[this.state.listExperimentsRequestId]}>
-        <HomeView experimentId={this.props.experimentId}/>
-      </RequestStateWrapper>
-    );
+    if (this.state.listExperimentsExpanded) {
+      return (
+        <div className="outer-container">
+          <div className="HomePage-experiment-list-container">
+            <RequestStateWrapper requestIds={[this.state.listExperimentsRequestId]}>
+              <div className="collapsed-expander-container">
+                <ExperimentListView
+                  activeExperimentId={this.props.experimentId}
+                  onClickListExperiments={this.onClickListExperiments}
+                />
+              </div>
+            </RequestStateWrapper>
+          </div>
+          <div className="experiment-view-container">
+            <ExperimentPage match={this.props.match}/>
+          </div>
+          <div className="experiment-view-right"/>
+        </div>
+      );
+    } else {
+      return (
+        <div>
+          <div className="collapsed-expander-container">
+            <i onClick={this.onClickListExperiments}
+               title="Show experiment list"
+               className="expander fa fa-chevron-right login-icon"/>
+          </div>
+          <div className="experiment-page-container">
+            <ExperimentPage match={this.props.match}/>
+          </div>
+        </div>
+      );
+    }
   }
 }
 
 const mapStateToProps = (state, ownProps) => {
   const { match } = ownProps;
   if (match.url === "/") {
-    return {};
+    return { experimentId: 0 };
   }
   return { experimentId: parseInt(match.params.experimentId, 10) };
 };
@@ -40,7 +80,7 @@ const mapStateToProps = (state, ownProps) => {
 const mapDispatchToProps = (dispatch) => {
   return {
     dispatchListExperimentsApi: (requestId) => {
-      return dispatch(listExperimentsApi(requestId));
+      dispatch(listExperimentsApi(requestId));
     }
   };
 };
