--- mlflow/server/js/src/components/NoExperimentView.js
@@ -1,9 +1,0 @@
-import React, { Component } from 'react';
-
-export default class NoExperimentView extends Component {
-  render() {
-    return (
-      <div>No experiment!</div>
-    );
-  }
-}
